refactor: change file names 'styles.tsx' to 'styled.tsx'
