+++ src/components/Layout/index.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router-dom";
 import Header from "../Header";
 import Navigation from "../Navigation";
-import * as S from "./styles";
+import * as S from "./styled";
 
 function Layout(): React.ReactNode {
   return (
